+++ MODULOS/FORNECEDORES.md
@@ -1,0 +1,20 @@
+# Módulo de Fornecedores
+
+Campos principais:
+- id
+- name
+- fantasy_name
+- email
+- phone
+- document
+- address
+- city
+- state
+- postal_code
+- contact_name
+- notes
+- is_active
+- created_at
+- updated_at
+
+Fluxo básico: listagem, criação, edição e exclusão de fornecedores com filtros avançados e exportação CSV.
